nixos/modules/xonsh/pyproject.toml
[tool.poetry]
name = "xonsh-extensions"
version = "0.1.0"
description = ""
authors = ["Jörg Thalheim <joerg@thalheim.io>"]

[tool.poetry.dependencies]
python = "^3.9"
xontrib-prompt-starship = { git = "https://github.com/anki-code/xontrib-prompt-starship" }
xontrib-fzf-widgets = { git = "https://github.com/Mic92/xontrib-fzf-widgets", branch = "patch-1" }
xonsh-direnv = { git = "https://github.com/74th/xonsh-direnv" }
xontrib-dracula = { git = "https://github.com/agoose77/xontrib-dracula", branch = "main" }
xontrib-sh = { git = "https://github.com/anki-code/xontrib-sh", branch = "main" }
xontrib-back2dir = { git = "https://github.com/anki-code/xontrib-back2dir" }
xontrib-output-search = { git = "https://github.com/tokenizer/xontrib-output-search" }

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
